Refactoring roles to use the new developer role and a role enumeration
