+++ src/main/java/org/sagebionetworks/bridge/sdk/Roles.java
@@ -1,0 +1,8 @@
+package org.sagebionetworks.bridge.sdk;
+
+public enum Roles {
+    DEVELOPER,
+    RESEARCHER,
+    ADMIN,
+    TEST_USERS
+}
